Successfully broke out all of my theme customizations into a separate subtheme so that we can (someday, maybe) upgrade the TB University, should an update become available.
This will provide for an easier, less damaging customization, as well. We'll be
able to work in a sandbox, rather than hacking up existing files, and possibly
losing unknown features in the process.

Best practice dictate that this should be done, and I'll most definitely
remember to take advantage of this kind of thing from the outset on my next
Drupal project.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>