Issue #26: Center the main menu on the screen
Added the CSS necessary to center the main menu on the screen.
I'm still not 100% sure if I like it that way, especially when
taken together with the breadcrumb directly underneath it. Let's
get some feedback.

We also should test this with mobile devices and across web
browsers, which may make working on this moot.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>