Issue #30: 404 Page Needs to Look Better
Changed the 'blackboard' image in the background, so that the owl
with the pointer is gone (why he was there to begin with, we'll
just never know...). Added the same logo that is in use in the header
on all pages across the site. Adjusted the wording in the HTML
template to be more grammatically correct.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>