Issue #17: Admissions Information Request
Fixed a peculiar behavior of (presumably all) our webforms when they
are loaded on a mobile device. The body of the form was displayed
with a top-border on it, and there was additional padding added to
the top, as well.

I've overridden these settings for webforms appearing on the frontpage,
since the only form that fits that is our short info request (this
may need to be adjusted down the road, if more forms make their way
onto the frontpage).

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>