Issue #30: 404 Page Needs to Look Better
At least for now, I've added a link to the search page, along with
the link to the home page. I think the "possibly related items"
would be a nice feature to have, but I don't want that to be a
show-stopper for us, this late in the game. We'll work on getting
that implemented later, but I don't want to tie up resources here
waiting to figure that out.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>