Issue #31: Added buttons to the top-right section in the header
I added two buttons to the top of each page: "Request Information"
and "Apply Now", along with the styling to format them in a nice
way (we still need to work on the floats and such in the responsive
layouts).

Below those buttons, I added three bits of information, that might
be helpful to site visitors: Phone number, a link to the "Contact
Us" page, and a link to a (yet to exist) FAQs page.

I wrapped up those buttons and links into a nice feature package, too.

During this process, I also installed and enabled the
contrib/fontawesome module (along with the required library) in
order to get my hands on some really nice icons.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>