Issue #18: Provided an emergency notification block
I set this up as a block that can be enabled or hidden
as needed, and will show on every page when it is enabled.

Right now, it's a "Warning" box, meaning that it's yellow
with a caution signe, and whatever text you want inside.

Overall, I'm very happy with how this turned out.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>