Issue #3: Finished off the Chapel Messages feature
I'm finally ready to say that we're done with this! I've got the
view looking good, and even taking a speaker name as a contextual
filter. The speaker names link to that contextual filter, and
provide only messages with that speaker. I even managed to get
the stupid breadcrumbs working the way that I wante to!

Also fixed a small issue where the "Contact Us" link in the
header was a relative URL.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>