Issue #3: Allowed even older chapel sermons to be created
For whatever reason, I had limited the age of a chapel sermon (year field,
at least) to one year ago. Don't know why I did that. For now, I've raised
that to the maximum of 10 years. Hopefully, I'll be able to figure out how
to raise it even more later on.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>