Issue #8: Implemented 'menu_block' for sidebar menus
The module seems to be working very well. I'm really happy with
the way that it's turned out so far. I was toying with the thought
of putting in some arrow indicators for menu items that have
children (like the superfish menu), but that looks like it's going
to be difficult to implement, and I wasn't altogether happy with
the results that I got initially. So, I'm not going to waste any
more time playing with it.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>