Re-creating some features, mostly due to block weights
Also changed the way the "read more" link is rendered in the
featured content view. The fa-circle-plus class is now applied
to the link, rather than adding it '::before' as I had done. So,
the fontawesome icon is now part of the link, which looks much
nicer, and behaves more intuitively.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>