Issue #1: Added the "auto_entitylabel" module to handle match titles
There's no real need (for us) for each athletic match to have a title.
Rather than delve into entity development and try to figure that all
out, this approach just allows me to set the title to specific tokens
and the people entering this stuff won't have to worry about it at all.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>