Issue #10: Style the student profile better, including social icons
If we want to include the social icons, the FontAwesome icons work very well
for that. However, it seems a little unnecessary to me.

I style the display better so that all the "bio information" goes on the
left, and the main chunch of text (questions and answers, currently) floats
to the right. It's a pretty nice effect is pretty nice, I think.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>