Issue #10: Replaced the age field with "year"
It's probably more useful (and more polite to the girls) to display
which year (freshman, sophomore, etc.) a student is than their
age. This will take a bit of updating, but at least that will only
have to be done once a year (after graduation), rather than a bunch
throughout the year.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>