Issue #39: Reversed the sort order for "features" to date, descending
Now that we're getting serious about adding some of that featured stuff,
it's time to reverse the sort order for good. The form (which will be
the smallest date, once we populate the rest of the content) will forever
remain on the far right (we might have to un-feature stuff as new stuff
comes up, though).

This will make our form look a little bit goofy for a little while, but
it will balance out once we create all the content that we need.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>