Issue #39: Added the correct image field to the basic page content type
As it turns out, the basic page had "field_img", rather than "field_image",
which was preventing the image from being displayed in the view. I don't
recall when or how this happened (if it was even something that I had done),
but I'm glad to have it sorted out now.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>