Fixed a typo in the last commit, added a custom module for our editor
The typo is simple, but it broke dev at one point, so it's best to get
it under control.

The custom module basically gives us a place to customize the config
of our CKEditor on the fly. The real reason for it is that it allows
me to also protect any <i class="fa-something"></i> tags, so that
our FontAwesome additions to pages won't disappear.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>