Updated the "read more" link in the "featured content" view
I didn't like that putting the FontAwesome CSS classes right in the
link turned the link font into a serif font, and it was way small!

The plus-circle icon is now a separate field and doesn't link to the
content, but the style fits better with the rest of the surrouding
context.

Signed-off-by: Elliot Voris <evoris@slcconline.edu>