Changes to the "degree program" and menu styling
I changed the sub-menu items to fit in more easily, while also
displaying the chevron-down that indicates it is a parent of more
content. This will make me feel better about displaying the pop-out
menus on those sub-items.

I moved the "How to apply" field to the bottom of the form, which
should make it easier to instruct others to ignore that item.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>