Issue #42: Created content type and views for the calendars
The content type is pretty simple, although I feel like it's a bit
more overt than I'd like for it to be. I couldn't figure out a way
in the view (and it might not be possible) to parse out whether or
not an item is in the fall or spring semester. Oh, well...

I'm also including the feed, too. It's pretty helpful, and I've got
a CSV that works wonders!

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>