Updated the social-links block to use FontAwesome and add more sites
I added a YouTube link, as well as a Pinterest link (which I had to
set up, and I'm probably not the best person for that job). We have
some decent starter content on both of those sites, but we'll need to
work toward having them become more regular parts of our social
interaction over time.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>