Reduced the number of chapel messages dispalyed
The mediaElement player was having trouble loading all the songs,
and it wouldn't load the last 14-or-so on a page of 20. Reducing
the number to 5 will help that.

It's a little more difficult, and I would (personally) rather see
all of them on one page if I could. But the search boxes eliminate
some of the hassle, at least.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>