Added the contrib/bracket module for the upcoming MCCC tournament
Coach has asked me to set up a bracket system for the tournament,
and I was really pleased to learn that there was already a module
that does almost exactly what I need.

I've worked on creating a bracket that models the one that coach
has supplied to me. Submodule is coming up next.

Signed-off-by: Elliot Voris <elliotfriend@gmail.com>