Issue #56: Sorting the degree_programs view by tid
I was a little nervous about having gotten lucky enough for the
blank sort of the view to work on its own. So, I wanted to overtly
sort the list somehow, into what we wanted.

Given that I've now completely exported the "Academic Program"
taxonomy terms into the "class" content type feature, the `tid` of
those particular terms are in exactly the order we want. So, now
we're going to sort by that value.

Easy peasy.

Signed-off-by: Elliot Voris <elliot@voris.me>