Issue #64: Fixing the Workbench Moderation dashboard views
At some point, we updated the workbench suite of modules, and it
looks like some changes were made to the way the view was created.

So, this has been a little messed up for a little while. Today,
we're fixing that!

Oh, and we're adding 5-minute cashing for query and output.

Signed-off-by: Elliot J. Voris <elliot@voris.me>